backend/requirements.txt
fastapi
uvicorn
langchain
chromadb
openai
python-dotenv 